Rearranged planscore module
