+++ Makefile
@@ -2,7 +2,7 @@ all: planscore-uploadfields.zip
 
 live: all
 	aws --region us-east-1 lambda update-function-code --function-name PlanScore-UploadFields --zip-file fileb://planscore-uploadfields.zip >> /dev/null
-	aws --region us-east-1 lambda update-function-configuration --function-name PlanScore-UploadFields --handler planscore.lambda_handler >> /dev/null
+	aws --region us-east-1 lambda update-function-configuration --function-name PlanScore-UploadFields --handler planscore.upload_fields.lambda_handler >> /dev/null
 
 planscore-uploadfields.zip:
 	mkdir -p planscore-uploadfields
